添加日志相关配置

+++ src/BiliLite.UWP/Services/LogHelper.cs
@@ -1,4 +1,5 @@
 ﻿using BiliLite.Extensions;
+using BiliLite.Helpers;
 using BiliLite.Models.Common;
 using NLog;
 using NLog.Config;
@@ -17,6 +18,10 @@ namespace BiliLite.Services
         public static LoggingConfiguration config;
         public static Logger logger = LogManager.GetCurrentClassLogger();
 
+        private static bool IsAutoClearLogFile => SettingHelper.GetValue<bool>(SettingHelper.Other.AUTO_CLEAR_LOG_FILE, true);
+        private static int AutoClearLogFileDay => SettingHelper.GetValue<int>(SettingHelper.Other.AUTO_CLEAR_LOG_FILE_DAY, 7);
+        private static bool IsProtectLogInfo => SettingHelper.GetValue<bool>(SettingHelper.Other.PROTECT_LOG_INFO, true);
+
         public static void Init()
         {
             config = new LoggingConfiguration();
@@ -33,14 +38,17 @@ namespace BiliLite.Services
             config.AddRule(LogLevel.Error, LogLevel.Error, logfile);
             config.AddRule(LogLevel.Fatal, LogLevel.Fatal, logfile);
             LogManager.Configuration = config;
-            // todo: add await
-            DeleteFile(storageFolder.Path + @"\log\");
+            if (IsAutoClearLogFile)
+            {
+                // todo: add await
+                DeleteFile(storageFolder.Path + @"\log\");
+            }
         }
 
         public static async Task DeleteFile(string path)
         {
             string pattern = "yyyyMMdd";
-            int days = 7;
+            int days = AutoClearLogFileDay;
             var folder = await StorageFolder.GetFolderFromPathAsync(path);
 
             var files = await folder.GetFilesAsync();
@@ -63,7 +71,8 @@ namespace BiliLite.Services
         {
             Debug.WriteLine("[" + LogType.INFO.ToString() + "]" + message);
             message = $"[{methodName}]{message}";
-            message = message.ProtectValues("access_key", "csrf", "access_token");
+            if(IsProtectLogInfo)
+                message = message.ProtectValues("access_key", "csrf", "access_token");
             switch (type)
             {
                 case LogType.INFO:
